Social Sustainability Global Database

The Social Sustainability global database and its visualization dashboard <https://public.tableau.com/app/profile/social.sustainability.and.inclusion.world.bank/viz/SocialSustainabilityGlobalDashboard2_0/Historia1?publish=yes/> are global public goods produced by the Social Development Global Practice of The World Bank Group. They feature leading indicators of inclusion, resilience, social cohesion, and process legitimacy, for 222 countries, disaggregated by population group and analyzed spatially and over time. In addition, the dashboard allows the user to overlay the indicators in the geospatial platform of the World Bank Group.

Usage

social_sustainability

Format

A data frame with 195 rows and 3 variables:

country_code

character World Bank country code.

year

numeric Year.

confidence_in_gov

numeric Percentage of population with confidence in government.

Source

World Bank https://datacatalog.worldbank.org/int/search/dataset/0061880/Social-Sustainability-Global-Database-